+++ ccd_portal/ccd_portal/doctype/ccd_portal_source_fixed_centre/ccd_portal_source_fixed_centre.py
@@ -1,0 +1,5 @@
+from frappe.model.document import Document
+
+
+class CCDPortalSourceFixedCentre(Document):
+	pass